requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
dataclasses==0.6
DateTime==4.3
idna==3.2
kaleido==0.2.1
multidict==5.2.0
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
plotly==5.3.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
slack-sdk==3.11.2
tenacity==8.0.1
twint @ git+https://github.com/twintproject/twint.git@4b30fa33991ccdf89c77471382daad3d3bc8c988
typing-extensions==3.10.0.2
urllib3==1.26.7
wincertstore==0.2
yarl==1.7.0
zope.interface==5.4.0
